[feature]: Parameter adjustment and adding community version prompts
