+++ packages/sqle/src/page/GlobalDashboard/List/PendingWorkOrder/index.tsx
@@ -45,7 +45,6 @@ const PendingWorkOrder: React.FC<GlobalDashboardListProps> = ({
           getGlobalWorkflowsV1FilterStatusListEnum.wait_for_audit,
           getGlobalWorkflowsV1FilterStatusListEnum.wait_for_execution,
           getGlobalWorkflowsV1FilterStatusListEnum.rejected,
-          getGlobalWorkflowsV1FilterStatusListEnum.canceled,
           getGlobalWorkflowsV1FilterStatusListEnum.exec_failed
         ]
       };
